Trait pallet_fragments_rpc::FragmentsRpcServer

pub trait FragmentsRpcServer<BlockHash, AccountId>: Sized + Send + Sync + 'static {
    fn get_definitions(
        &self,
        params: GetDefinitionsParams<AccountId, String>,
        at: Option<BlockHash>
    ) -> RpcResult<String>;
    fn get_instances(
        &self,
        params: GetInstancesParams<AccountId, String>,
        at: Option<BlockHash>
    ) -> RpcResult<String>;
    fn get_instance_owner(
        &self,
        params: GetInstanceOwnerParams<String>,
        at: Option<BlockHash>
    ) -> RpcResult<String>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        AccountId: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the FragmentsRpc RPC API.

Required Methods

fn get_definitions(
&self,
params: GetDefinitionsParams<AccountId, String>,
at: Option<BlockHash>
) -> RpcResult<String>

Query and Return Fragment Definition(s) based on params

fn get_instances(
&self,
params: GetInstancesParams<AccountId, String>,
at: Option<BlockHash>
) -> RpcResult<String>

Query and Return Fragment Instance(s) based on params

fn get_instance_owner(
&self,
params: GetInstanceOwnerParams<String>,
at: Option<BlockHash>
) -> RpcResult<String>

Query the owner of a Fragment Instance. The return type is a String

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
BlockHash: Send + Sync + 'static + DeserializeOwned,
AccountId: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.